% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim_SI.R
\name{clust_sim_SI}
\alias{clust_sim_SI}
\title{clust_sim_SI}
\usage{
clust_sim_SI(network_el, beta, intxn_per_day, days, iters, free_threads = 1)
}
\arguments{
\item{free_threads}{}
}
\description{
clust_sim_SI
}
