% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim_SI_unif.R
\name{clust_sim_SI_unif}
\alias{clust_sim_SI_unif}
\title{clust_sim_SI_unif}
\usage{
clust_sim_SI_unif(network_el, beta, days, iters, free_threads = 1)
}
\arguments{
\item{network_el}{an edgelist file generated by igraph}

\item{beta}{a value from 0 to 1 indicating per contract probability of
pathogen transmission}

\item{days}{the maximum number of days for which your simulation will run}

\item{iters}{the number of times to replicate simulations}

\item{free_threads}{the number of computing threads to be left unused by
simulations (e.g. for doing other tasks while simulations are running.)}
}
\value{
a matrix with the same number of columns as argument 'iters', where
  each column represents an iteration, and each row represents 1: number of
  days simulation went for, 2: final number of susceptible individuals,
  3: final number of infected individuals, in that order.
}
\description{
clust_sim_SI_unif simulates the spread of an infectious disease on a social
network using the Susceptible-Infected compartmental model with one uniform
interaction per dyad per day in parallel for a given number of iterations.
}
