% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim_SIR.R
\name{clust_sim_SIR}
\alias{clust_sim_SIR}
\title{clust_sim_SIR}
\usage{
clust_sim_SIR(network_el, beta, gamma, intxn_per_day, days, iters,
  free_threads = 1)
}
\arguments{
\item{free_threads}{}
}
\description{
clust_sim_SIR
}
