% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_SI_erg_ens.R
\name{compare_SI_erg_ens}
\alias{compare_SI_erg_ens}
\title{compare_SI_erg_ens}
\usage{
compare_SI_erg_ens(sim_erg_output, sim_max_output, n_min)
}
\arguments{
\item{n_min}{}
}
\description{
compare_SI_erg_ens
}
