% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_max.R
\name{gen_max}
\alias{gen_max}
\title{gen_max}
\usage{
gen_max(n_min, n_max)
}
\arguments{
\item{n_min}{An integer of the lowest number of nodes to be generated as a maximally
complete graph}

\item{n_max}{An integer of the highest number of nodes to be generated as a maximally
complete graph}
}
\value{
list wirh 2 elements, the first of which (g) contains sociomatrices for each
  maximally complete graph, and the second of which (el) contains edgelists for those
  same graphs
}
\description{
gen_max generates a set of maximally complete graphs from size "n_min" to "n_max",
  provided in the function call
}
