% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_SI.R
\name{plot_compare_SI}
\alias{plot_compare_SI}
\title{plot_compare_SI}
\usage{
plot_compare_SI(sim_erg_output, sim_max_output, n_min, n, p, samp)
}
\arguments{
\item{samp}{}
}
\description{
plot_compare_SI
}
