% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_SIR.R
\name{plot_compare_SIR}
\alias{plot_compare_SIR}
\title{plot_compare_SIR}
\usage{
plot_compare_SIR(sim_erg_output, sim_max_output, max_nmin, n, p, samp)
}
\arguments{
\item{samp}{}
}
\description{
plot_compare_SIR
}
