% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SI.R
\name{sim_SI}
\alias{sim_SI}
\title{sim_SI}
\usage{
sim_SI(network_el, beta, intxn_per_day, days)
}
\arguments{
\item{network_el}{an edgelist file generated by igraph}

\item{beta}{a value from 0 to 1 indicating per contract probability of
pathogen transmission}

\item{intxn_per_day}{the per capita interaction rate}

\item{days}{the maximum number of days for which your simulation will run}
}
\value{
vector with 3 elements, 1: number of days simulation went for, 2:
  final number of susceptible individuals, 3: final numbre of infected
  individuals
}
\description{
sim_SI simulates the spread of an infectious disease on a social network
using the Susceptible-Infected compartmental model
}
