% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SI_unif.R
\name{sim_SI_unif}
\alias{sim_SI_unif}
\title{sim_SI_unif}
\usage{
sim_SI_unif(network_el, beta, days)
}
\arguments{
\item{network_el}{an edgelist file generated by igraph}

\item{beta}{a value from 0 to 1 indicating per contract probability of
pathogen transmission}

\item{days}{the maximum number of days for which your simulation will run}
}
\value{
vector with 3 elements, 1: number of days simulation went for, 2:
  final number of susceptible individuals, 3: final numbre of infected
  individuals
}
\description{
sim_SI_unif simulates the spread of an infectious disease on a social network
using the Susceptible-Infected compartmental model with one uniform interaction
per dyad per day
}
