% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SIR.R
\name{sim_SIR}
\alias{sim_SIR}
\title{sim_SIR}
\usage{
sim_SIR(network_el, beta, gamma, intxn_per_day, days)
}
\arguments{
\item{network_el}{an edgelist file generated by igraph}

\item{beta}{a value from 0 to 1 indicating per contract probability of
pathogen transmission}

\item{gamma}{a value from 0 to 1 indicating daily probability of
recovery/removal}

\item{intxn_per_day}{the per capita interaction rate}

\item{days}{the maximum number of days for which your simulation will run}
}
\value{
vector with 5 elements, 1: number of days simulation went for, 2:
  final number of susceptible individuals, 3: final number of infected
  individuals, 4: final number of number of removed individuals, 5: maximum
  number of infected inviduals at any timepoint
}
\description{
sim_SIR simulates the spread of an infectious disease on a social network
using the Susceptible-Infected-Removed compartmental model
}
