% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_SIR_unif.R
\name{sim_SIR_unif}
\alias{sim_SIR_unif}
\title{sim_SIR_unif}
\usage{
sim_SIR_unif(network_el, beta, gamma, days)
}
\arguments{
\item{network_el}{an edgelist file generated by igraph}

\item{beta}{a value from 0 to 1 indicating per contract probability of
pathogen transmission}

\item{gamma}{a value from 0 to 1 indicating daily probability of
recovery/removal}

\item{days}{the maximum number of days for which your simulation will run}
}
\value{
vector with 5 elements, 1: number of days simulation went for, 2:
  final number of susceptible individuals, 3: final number of infected
  individuals, 4: final number of number of removed individuals, 5: maximum
  number of infected inviduals at any timepoint
}
\description{
sim_SIR_unif simulates the spread of an infectious disease on a social network
using the Susceptible-Infected-Removed compartmental model with one uniform
interaction per dyad per day
}
