% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_sim_SIR_unif.R
\name{clust_sim_SIR_unif}
\alias{clust_sim_SIR_unif}
\title{clust_sim_SIR_unif}
\usage{
clust_sim_SIR_unif(network_el, beta, gamma, days, iters, free_threads = 1)
}
\arguments{
\item{free_threads}{}
}
\description{
clust_sim_SIR_unif
}
