% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_backtrans_ens.R
\name{estimate_backtrans_ens}
\alias{estimate_backtrans_ens}
\title{estimate_backtrans_ens}
\usage{
estimate_backtrans_ens(x, predict_output, sim_output)
}
\arguments{
\item{sim_output}{}
}
\description{
estimate_backtrans_ens
}
