% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ens.R
\name{estimate_ens}
\alias{estimate_ens}
\title{estimate_ens}
\usage{
estimate_ens(x, predict_model)
}
\arguments{
\item{x}{the number of days that a disease took to break out through the observed network}

\item{predict_model}{model output from either 'predict_SI_max' or 'predict_SIR_max'}
}
\value{
integer predicted effective network size
}
\description{
estimate_ens takes the outbreak time for a given network and compares this to the
  relationship between group size and outbreak time on maximally to predict
  effective network size
}
